move instrumentatino logic out of script.ts
